@@ -1,42 +1,48 @@
-from __future__ import division, print_function
+import re
+
 import pytest
 
+from humre import (ALPHANUMERIC, ANYCHAR, ANYTHING, ASTERISK, CLOSE_PAREN,
+                   DIGIT, EVERYTHING, OPEN_PAREN, PERIOD, at_least, at_most,
+                   between, compile, ends_with, exactly, group, group_chars,
+                   group_either, one_or_more, starts_and_ends_with,
+                   chars, WHITESPACE,
+                   starts_with, zero_or_more)
 from humre import *
-import re, sys
 
 
-def test_random_regexes_I_found_online():
+def test_random_regexes_I_found_online() -> None:
     # Basic American phone number regex:
     assert (
-        r'\d\d\d-\d\d\d-\d\d\d\d'
-        == DIGIT + DIGIT + DIGIT + '-' + DIGIT + DIGIT + DIGIT + '-' + DIGIT + DIGIT + DIGIT + DIGIT
+        DIGIT + DIGIT + DIGIT + '-' + DIGIT + DIGIT + DIGIT + '-' + DIGIT + DIGIT + DIGIT + DIGIT
+        == r'\d\d\d-\d\d\d-\d\d\d\d'
     )
 
     # Basic American phone number regex using {3}:
     # assert r'\d{3}-\d{3}-\d{4}' == exactly_3(DIGIT) + '-' + exactly_3(DIGIT) + '-' + exactly_4(DIGIT)
-    assert r'\d{3}-\d{3}-\d{4}' == exactly(3, DIGIT) + '-' + exactly(3, DIGIT) + '-' + exactly(4, DIGIT)
+    assert exactly(3, DIGIT) + '-' + exactly(3, DIGIT) + '-' + exactly(4, DIGIT) == r'\d{3}-\d{3}-\d{4}'
 
     # American phone number with groups:
-    assert r'(\d{3})-(\d{3}-\d{4})' == group(exactly(3, DIGIT)) + '-' + group(
+    assert group(exactly(3, DIGIT)) + '-' + group(
         exactly(3, DIGIT) + '-' + exactly(4, DIGIT)
-    )
+    ) == r'(\d{3})-(\d{3}-\d{4})'
 
-    assert 'First Name: (.*?)' == 'First Name: ' + group(ANYTHING)
-    assert 'First Name: (.*)' == 'First Name: ' + group(EVERYTHING)
+    assert 'First Name: ' + group(ANYTHING) == 'First Name: (.*?)'
+    assert 'First Name: ' + group(EVERYTHING) == 'First Name: (.*)'
 
-    assert 'x*' == zero_or_more('x')
-    assert 'x+' == one_or_more('x')
+    assert zero_or_more('x') == 'x*'
+    assert one_or_more('x') == 'x+'
 
-    assert 'x{3,5}' == between(3, 5, 'x')
-    assert 'x{2,}' == at_least(2, 'x')
+    assert between(3, 5, 'x') == 'x{3,5}'
+    assert at_least(2, 'x') == 'x{2,}'
 
-    assert 'x{,2}' == at_most(2, 'x')
+    assert at_most(2, 'x') == 'x{,2}'
 
-    assert '^x' == starts_with('x')
-    assert 'x$' == ends_with('x')
-    assert '^x$' == starts_and_ends_with('x')
+    assert starts_with('x') == '^x'
+    assert ends_with('x') == 'x$'
+    assert starts_and_ends_with('x') == '^x$'
 
-    assert r'((\d{3}|\(\d{3}\))?(\s|-|\.)?\d{3}(\s|-|\.)\d{4}(\s*(ext|x|ext.)\s*\d{2,5})?)' == group(
+    assert group(
         # Area code:
         optional(group(either(exactly(3, DIGIT), OPEN_PAREN + exactly(3, DIGIT) + CLOSE_PAREN)))
         + optional(group(either(WHITESPACE, '-', PERIOD)))
@@ -57,10 +63,10 @@ def test_random_regexes_I_found_online():
                 + between(2, 5, DIGIT)
             )
         )
-    )
+    ) == r'((\d{3}|\(\d{3}\))?(\s|-|\.)?\d{3}(\s|-|\.)\d{4}(\s*(ext|x|ext.)\s*\d{2,5})?)'
 
     # Use commas instead of + str concatenation:
-    assert r'((\d{3}|\(\d{3}\))?(\s|-|\.)?\d{3}(\s|-|\.)\d{4}(\s*(ext|x|ext.)\s*\d{2,5})?)' == group(
+    assert group(
         optional(group(either(exactly(3, DIGIT), OPEN_PAREN + exactly(3, DIGIT) + CLOSE_PAREN))),
         optional(group(either(WHITESPACE, '-', PERIOD))),
         exactly(3, DIGIT),
@@ -74,9 +80,9 @@ def test_random_regexes_I_found_online():
                 between(2, 5, DIGIT),
             )
         ),
-    )
+    ) == r'((\d{3}|\(\d{3}\))?(\s|-|\.)?\d{3}(\s|-|\.)\d{4}(\s*(ext|x|ext.)\s*\d{2,5})?)'
 
-    assert r'((\d{3}|\(\d{3}\))?(\s|-|\.)?\d{3}(\s|-|\.)\d{4}(\s*(ext|x|ext.)\s*\d{2,5})?)' == group(
+    assert group(
         optional_group(either(exactly(3, DIGIT), r'\(' + exactly(3, DIGIT) + r'\)'))
         + optional_group(either(WHITESPACE, '-', PERIOD))
         + exactly(3, DIGIT)
@@ -88,22 +94,22 @@ def test_random_regexes_I_found_online():
             + zero_or_more(WHITESPACE)
             + between(2, 5, DIGIT)
         )
-    )
+    ) == r'((\d{3}|\(\d{3}\))?(\s|-|\.)?\d{3}(\s|-|\.)\d{4}(\s*(ext|x|ext.)\s*\d{2,5})?)'
 
-    assert '[YZ][BCE-HMO-Y][BEFN][A-Z][0-9][0-9]_KWBC_[0-9]{6}' == chars('YZ') + chars('BCE-HMO-Y') + chars(
-        'BEFN'
-    ) + chars('A-Z') + chars('0-9') + chars('0-9') + '_KWBC_' + exactly(6, chars('0-9'))
-    assert r'\s*\\\\$' == ends_with(zero_or_more(WHITESPACE) + '\\\\\\\\')
+    assert (chars('YZ') + chars('BCE-HMO-Y') + chars('BEFN') + chars('A-Z') + chars('0-9') +
+            chars('0-9') + '_KWBC_' + exactly(6, chars('0-9'))
+            == '[YZ][BCE-HMO-Y][BEFN][A-Z][0-9][0-9]_KWBC_[0-9]{6}')
+    assert ends_with(zero_or_more(WHITESPACE) + '\\\\\\\\') == r'\s*\\\\$'
 
-    assert r"^Access:" == starts_with('Access:')
+    assert starts_with('Access:') == r"^Access:"
 
     import string
 
-    assert "Check[ ]?sum[ ]+is[ ]+([{}])".format(string.printable) == 'Check' + optional(
+    assert f"Check[ ]?sum[ ]+is[ ]+([{string.printable}])" == 'Check' + optional(
         chars(' ')
     ) + 'sum' + one_or_more(chars(' ')) + 'is' + one_or_more(chars(' ')) + group(chars(string.printable))
     # Use ''.join() so that I can have commas instead.
-    assert "Check[ ]?sum[ ]+is[ ]+([{}])".format(string.printable) == ''.join(
+    assert f"Check[ ]?sum[ ]+is[ ]+([{string.printable}])" == ''.join(
         (
             # The WORD 'Checksum' with optional space between 'check' and 'sum':
             'Check',
@@ -120,103 +126,103 @@ def test_random_regexes_I_found_online():
         )
     )
 
-    assert r'[-]?\d\n' == optional(chars('-')) + DIGIT + NEWLINE
+    assert optional(chars('-')) + DIGIT + NEWLINE == r'[-]?\d\n'
 
-    assert r'Student has an ([A-D]) grade' == 'Student has an ' + group(chars('A-D')) + ' grade'
+    assert 'Student has an ' + group(chars('A-D')) + ' grade' == r'Student has an ([A-D]) grade'
 
-    assert r'([-]?\d+) is the median' == group(optional(chars('-')), one_or_more(DIGIT)) + ' is the median'
+    assert group(optional(chars('-')), one_or_more(DIGIT)) + ' is the median' == r'([-]?\d+) is the median'
 
-    assert '^[A-Z0-9]' == starts_with(chars('A-Z0-9'))
+    assert starts_with(chars('A-Z0-9')) == '^[A-Z0-9]'
 
-    assert r"\s(\S):\s" == WHITESPACE + group(NONWHITESPACE) + ':' + WHITESPACE
-    assert r"\s(\S):\s" == ''.join([WHITESPACE, group(NONWHITESPACE), ':', WHITESPACE])
+    assert WHITESPACE + group(NONWHITESPACE) + ':' + WHITESPACE == r"\s(\S):\s"
+    assert ''.join([WHITESPACE, group(NONWHITESPACE), ':', WHITESPACE]) == r"\s(\S):\s"
     assert re.compile(r"\s(\S):\s") == compile(WHITESPACE, group(NONWHITESPACE), ':', WHITESPACE)
     assert re.compile(r"\s(\S):\s") == re.compile(''.join((WHITESPACE, group(NONWHITESPACE), ':', WHITESPACE)))
 
-    assert r"\*([^\*]+)\*" == ASTERISK + group(one_or_more(nonchars(ASTERISK))) + ASTERISK
+    assert ASTERISK + group(one_or_more(nonchars(ASTERISK))) + ASTERISK == r"\*([^\*]+)\*"
 
     assert zero_or_more(BACKSLASH) == r'\\*'
     assert ASTERISK == r'\*'
 
-    assert r"\s+" == one_or_more(WHITESPACE)
+    assert one_or_more(WHITESPACE) == r"\s+"
 
-    assert "^yslogin[0-9]+" == starts_with('yslogin', one_or_more(chars('0-9')))
+    assert starts_with('yslogin', one_or_more(chars('0-9'))) == "^yslogin[0-9]+"
 
     assert (
-        r"\$\((?P<name>[A-Za-z0-9_]+)\)"
-        == DOLLAR + OPEN_PAREN + named_group('name', one_or_more(chars('A-Za-z0-9_'))) + CLOSE_PAREN
+        DOLLAR + OPEN_PAREN + named_group('name', one_or_more(chars('A-Za-z0-9_'))) + CLOSE_PAREN
+        == r"\$\((?P<name>[A-Za-z0-9_]+)\)"
     )
 
-    assert "ccd id: \\d+" == 'ccd id: ' + one_or_more(DIGIT)
+    assert 'ccd id: ' + one_or_more(DIGIT) == "ccd id: \\d+"
 
     # assert "[%:\r\n]" == chars(r'%:\r\n')  # Test this: is there a difference between "[%:\r\n]" and r"[%:\r\n]"
 
-    assert "%([0-9a-fA-F][0-9a-fA-F])" == '%' + group(chars('0-9a-fA-F'), chars('0-9a-fA-F'))
+    assert '%' + group(chars('0-9a-fA-F'), chars('0-9a-fA-F')) == "%([0-9a-fA-F][0-9a-fA-F])"
 
-    assert r'\?sid=(?P<sid>\d+)' == QUESTION_MARK + 'sid=' + named_group('sid', one_or_more(DIGIT))
+    assert QUESTION_MARK + 'sid=' + named_group('sid', one_or_more(DIGIT)) == r'\?sid=(?P<sid>\d+)'
 
     assert (
-        'INPUT.*NAME="seui".*VALUE="(?P<uid>[^"]*)"'
-        == 'INPUT'
+        'INPUT'
         + EVERYTHING
         + 'NAME="seui"'
         + EVERYTHING
         + 'VALUE="'
         + named_group('uid', zero_or_more(nonchars('"')))
         + '"'
+        == 'INPUT.*NAME="seui".*VALUE="(?P<uid>[^"]*)"'
     )
 
     assert (
-        'INPUT.*NAME="sdn".*VALUE="(?P<rnm>[^"]*)"'
-        == 'INPUT'
+        'INPUT'
         + EVERYTHING
         + 'NAME="sdn"'
         + EVERYTHING
         + 'VALUE="'
         + named_group('rnm', zero_or_more(nonchars('"')))
         + '"'
+        == 'INPUT.*NAME="sdn".*VALUE="(?P<rnm>[^"]*)"'
     )
 
     assert (
-        'OPTION VALUE="(?P<mod>[^"]*)" SELECTED'
-        == 'OPTION VALUE="' + named_group('mod', zero_or_more(nonchars('"'))) + '" SELECTED'
+        'OPTION VALUE="' + named_group('mod', zero_or_more(nonchars('"'))) + '" SELECTED'
+        == 'OPTION VALUE="(?P<mod>[^"]*)" SELECTED'
     )
 
     assert (
-        'INPUT.*NAME="sena" VALUE="(?P<ena>[^"]*)" CHECKED'
-        == 'INPUT' + EVERYTHING + 'NAME="sena" VALUE="' + named_group('ena', zero_or_more(nonchars('"'))) + '" CHECKED'
+        'INPUT' + EVERYTHING + 'NAME="sena" VALUE="' + named_group('ena', zero_or_more(nonchars('"'))) + '" CHECKED'
+        == 'INPUT.*NAME="sena" VALUE="(?P<ena>[^"]*)" CHECKED'
     )
 
-    assert r"[.?!]" == chars('.?!')  # Functionally equivalent to chars(PERIOD + QUESTION_MARK + '!')
+    assert chars('.?!') == r"[.?!]"  # Functionally equivalent to chars(PERIOD + QUESTION_MARK + '!')
 
-    assert r"[A-Z]\w+" == chars('A-Z') + one_or_more(WORD)
+    assert chars('A-Z') + one_or_more(WORD) == r"[A-Z]\w+"
 
     # This has a bug in it because the . dots can match ANYTHING:
-    assert r'^pip-.*(zip|tar.gz|tar.bz2|tgz|tbz)$' == starts_and_ends_with(
+    assert starts_and_ends_with(
         'pip-' + EVERYTHING + group_either('zip', 'tar.gz', 'tar.bz2', 'tgz', 'tbz')
-    )
+    ) == r'^pip-.*(zip|tar.gz|tar.bz2|tgz|tbz)$'
 
-    assert r'/Python(?:-32|-64)*$' == '/Python' + ends_with(zero_or_more(noncap_group(either('-32', '-64'))))
+    assert '/Python' + ends_with(zero_or_more(noncap_group(either('-32', '-64')))) == r'/Python(?:-32|-64)*$'
 
-    assert r'^CPU\(s\):\s*(\d+)' == starts_with(
+    assert starts_with(
         'CPU' + OPEN_PAREN + 's' + CLOSE_PAREN + ':' + zero_or_more(WHITESPACE)
-    ) + group(one_or_more(DIGIT))
+    ) + group(one_or_more(DIGIT)) == r'^CPU\(s\):\s*(\d+)'
 
-    assert '^GOOGLE_RELEASE=(.+)$' == starts_and_ends_with('GOOGLE_RELEASE=', group(one_or_more(ANYCHAR)))
+    assert starts_and_ends_with('GOOGLE_RELEASE=', group(one_or_more(ANYCHAR))) == '^GOOGLE_RELEASE=(.+)$'
 
-    assert '^(.+)=(.+)$' == starts_and_ends_with(group(one_or_more(ANYCHAR)), '=', group(one_or_more(ANYCHAR)))
+    assert starts_and_ends_with(group(one_or_more(ANYCHAR)), '=', group(one_or_more(ANYCHAR))) == '^(.+)=(.+)$'
 
-    assert '<.*?>' == '<' + ANYTHING + '>'
+    assert '<' + ANYTHING + '>' == '<.*?>'
 
-    assert r'chr(\d+):(\d+)(-(\d+))?' == 'chr' + group(one_or_more(DIGIT)) + ':' + group(
+    assert 'chr' + group(one_or_more(DIGIT)) + ':' + group(
         one_or_more(DIGIT)
-    ) + optional_group('-', group(one_or_more(DIGIT)))
+    ) + optional_group('-', group(one_or_more(DIGIT))) == r'chr(\d+):(\d+)(-(\d+))?'
 
-    assert r'chr(\d+)(:(\d+)(-(\d+))?)?' == 'chr' + group(one_or_more(DIGIT)) + optional_group(
+    assert 'chr' + group(one_or_more(DIGIT)) + optional_group(
         ':', group(one_or_more(DIGIT)), optional_group('-', group(one_or_more(DIGIT)))
-    )
+    ) == r'chr(\d+)(:(\d+)(-(\d+))?)?'
 
-    assert "[^a-zA-Z0-9']+" == one_or_more(nonchars("a-zA-Z0-9'"))
+    assert one_or_more(nonchars("a-zA-Z0-9'")) == "[^a-zA-Z0-9']+"
 
     # Purposefully left this as a non-raw string:
     assert re.compile(r"<person>([,\s]*(and)*[,\s]*<person>)+") == compile(
@@ -229,7 +235,7 @@ def test_random_regexes_I_found_online():
     # assert re.compile(r"[()[\].,|:;?!=+~\-\/{}]") == compile(chars('()[\].,|:;?!=+~\-\/{}'))
     assert re.compile(r"[()[].,|:;?!=+~-/{}]") == compile(chars('()[].,|:;?!=+~-/{}'))
 
-    assert re.compile('''['"`]''') == compile(chars("""'"`"""))
+    assert re.compile("""['"`]""") == compile(chars("""'"`"""))
 
     assert re.compile(r'(\s*"+\s*)+') == compile(
         one_or_more_group(zero_or_more(WHITESPACE), one_or_more('"'), zero_or_more(WHITESPACE))
@@ -237,14 +243,13 @@ def test_random_regexes_I_found_online():
 
     assert re.compile(r"(\d),(\d{3})") == compile(group(DIGIT), ',', group(exactly(3, DIGIT)))
 
-    assert r"(\w)\.(\w)" == group(WORD) + PERIOD + group(WORD)
+    assert group(WORD) + PERIOD + group(WORD) == r"(\w)\.(\w)"
 
     # TODO - research: positive lookahead and negative lookahead (?=
     # assert r"'(?=([stdm]|(ll)|(re)|(ve)|(ll))\b)" == "'" + group()
 
     assert (
-        r'<math.*?alttext="\{\\displaystyle (.*?)\}"'
-        == '<math'
+        '<math'
         + ANYTHING
         + 'alttext="'
         + OPEN_BRACE
@@ -253,43 +258,44 @@ def test_random_regexes_I_found_online():
         + group(ANYTHING)
         + CLOSE_BRACE
         + '"'
+        == r'<math.*?alttext="\{\\displaystyle (.*?)\}"'
     )
 
-    assert r'([^-+.:0-9])+' == one_or_more_group(nonchars('-+.:0-9'))
+    assert one_or_more_group(nonchars('-+.:0-9')) == r'([^-+.:0-9])+'
 
-    assert r'(\d+)' == group(one_or_more(DIGIT))
+    assert group(one_or_more(DIGIT)) == r'(\d+)'
 
-    assert r"(\w+(\W*\d+\W*)?\-?\w*?[^\S\t]*)*" == zero_or_more_group(
+    assert zero_or_more_group(
         one_or_more(WORD),
         optional_group(zero_or_more(NONWORD), one_or_more(DIGIT), zero_or_more(NONWORD)),
         optional(r'\-'),
         optional(zero_or_more(WORD)),
         zero_or_more(nonchars(NONWHITESPACE, TAB)),
-    )
+    ) == r"(\w+(\W*\d+\W*)?\-?\w*?[^\S\t]*)*"
 
-    assert r"(\w+(\W*\d+\W*)?\-?\w*?[^\S\t]*)*" == zero_or_more_group(
+    assert zero_or_more_group(
         one_or_more(WORD),
         optional_group(zero_or_more(NONWORD), one_or_more(DIGIT), zero_or_more(NONWORD)),
         optional('\\-'),
         optional(zero_or_more(WORD)),
         zero_or_more(nonchars(NONWHITESPACE, TAB)),
-    )
+    ) == r"(\w+(\W*\d+\W*)?\-?\w*?[^\S\t]*)*"
 
-    assert r"'(?=([stdm]|(ll)|(re)|(ve)|(ll))\b)" == "'" + positive_lookahead(
+    assert "'" + positive_lookahead(
         group_either(chars('stdm'), group('ll'), group('re'), group('ve'), group('ll')), BOUNDARY
-    )
+    ) == r"'(?=([stdm]|(ll)|(re)|(ve)|(ll))\b)"
 
-    assert r"(\s*,+\s*)+" == one_or_more_group(zero_or_more(WHITESPACE), one_or_more(','), zero_or_more(WHITESPACE))
+    assert one_or_more_group(zero_or_more(WHITESPACE), one_or_more(','), zero_or_more(WHITESPACE)) == r"(\s*,+\s*)+"
 
-    assert r'[\W_]' == chars(NONWORD, '_')
+    assert chars(NONWORD, '_') == r'[\W_]'
 
 
-def test_join():
-    assert join('dog', 'cat', 'moose') == ''.join(['dog', 'cat', 'moose'])
+def test_join() -> None:
+    assert join('dog', 'cat', 'moose') == 'dogcatmoose'
     assert join() == ''
 
 
-def test_esc():
+def test_esc() -> None:
     assert esc(r'hello') == re.escape(r'hello') == 'hello'
 
     for char in '.^$*+?{}[]\\|()':
@@ -298,34 +304,34 @@ def test_esc():
     assert esc(r'+') == re.escape(r'+') == r'\+'
 
 
-def test_compile():
+def test_compile() -> None:
     assert compile('hello') == re.compile('hello')
     assert compile('hello', IGNORECASE=True, DOTALL=True) == re.compile('hello', re.IGNORECASE | re.DOTALL)
 
-    assert compile('hello', A=True) == re.compile('hello', re.A)
+    assert compile('hello', A=True) == re.compile('hello', re.ASCII)
     assert compile('hello', ASCII=True) == re.compile('hello', re.ASCII)
     assert compile('hello', DEBUG=True) == re.compile('hello', re.DEBUG)
-    assert compile('hello', I=True) == re.compile('hello', re.I)
+    assert compile('hello', I=True) == re.compile('hello', re.IGNORECASE)
     assert compile('hello', IGNORECASE=True) == re.compile('hello', re.IGNORECASE)
     # Humre is not going to support bytes objects, which LOCALE requires.
     # assert compile(b'hello', L=True) == re.compile(b'hello', re.L)
     # assert compile(b'hello', LOCALE=True) == re.compile(b'hello', re.LOCALE)
-    assert compile('hello', M=True) == re.compile('hello', re.M)
+    assert compile('hello', M=True) == re.compile('hello', re.MULTILINE)
     assert compile('hello', MULTILINE=True) == re.compile('hello', re.MULTILINE)
-    assert compile('hello', S=True) == re.compile('hello', re.S)
+    assert compile('hello', S=True) == re.compile('hello', re.DOTALL)
     assert compile('hello', DOTALL=True) == re.compile('hello', re.DOTALL)
-    assert compile('hello', X=True) == re.compile('hello', re.X)
+    assert compile('hello', X=True) == re.compile('hello', re.VERBOSE)
     assert compile('hello', VERBOSE=True) == re.compile('hello', re.VERBOSE)
 
 
-def test_group():
+def test_group() -> None:
     assert group('cat') == '(cat)'
     assert group('cat', 'dog', 'moose') == '(catdogmoose)'
     assert group('cat', group('dog', 'moose')) == '(cat(dogmoose))'
     assert group('cat', group('dog', group('moose'))) == '(cat(dog(moose)))'
 
 
-def test_positive_lookahead():
+def test_positive_lookahead() -> None:
     assert lookahead('cat') == '(?=cat)'
     assert lookahead('cat', 'dog', 'moose') == '(?=catdogmoose)'
 
@@ -333,12 +339,12 @@ def test_positive_lookahead():
     assert positive_lookahead('cat', 'dog', 'moose') == '(?=catdogmoose)'
 
 
-def test_negative_lookahead():
+def test_negative_lookahead() -> None:
     assert negative_lookahead('cat') == '(?!cat)'
     assert negative_lookahead('cat', 'dog', 'moose') == '(?!catdogmoose)'
 
 
-def test_positive_lookbehind():
+def test_positive_lookbehind() -> None:
     assert lookbehind('cat') == '(?<=cat)'
     assert lookbehind('cat', 'dog', 'moose') == '(?<=catdogmoose)'
 
@@ -346,34 +352,34 @@ def test_positive_lookbehind():
     assert positive_lookbehind('cat', 'dog', 'moose') == '(?<=catdogmoose)'
 
 
-def test_negative_lookbehind():
+def test_negative_lookbehind() -> None:
     assert negative_lookbehind('cat') == '(?<!cat)'
     assert negative_lookbehind('cat', 'dog', 'moose') == '(?<!catdogmoose)'
 
 
-def test_named_group():
-    with pytest.raises(ValueError) as excObj:
+def test_named_group() -> None:
+    with pytest.raises(ValueError):
         named_group('', 'hello')  # Blank name.
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         named_group('2', 'hello')  # Starts with number.
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         named_group('!', 'hello')  # Invalid character.
 
     assert named_group('foo', 'cat') == '(?P<foo>cat)'
     assert named_group('foo', 'cat', 'dog', 'moose') == '(?P<foo>catdogmoose)'
 
 
-def test_noncap_group():
+def test_noncap_group() -> None:
     assert noncap_group('cat') == '(?:cat)'
     assert noncap_group('cat', 'dog', 'moose') == '(?:catdogmoose)'
 
 
-def test_optional():
-    with pytest.raises(ValueError) as excObj:
+def test_optional() -> None:
+    with pytest.raises(ValueError):
         optional()  # No args.
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         optional('')  # Blank arg.
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         optional('', '')  # Blank args.
 
     assert optional('c') == 'c?'
@@ -382,12 +388,12 @@ def test_optional():
     assert group(optional('cat')) == '(cat?)'
 
 
-def test_either():
-    with pytest.raises(ValueError) as excObj:
+def test_either() -> None:
+    with pytest.raises(ValueError):
         either()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         either('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         either('', '')
 
     assert either('cat', 'dog', 'moose') == 'cat|dog|moose'
@@ -395,18 +401,18 @@ def test_either():
     assert group(either('cat', 'dog', 'moose')) == '(cat|dog|moose)'
 
 
-def test_exactly():
-    with pytest.raises(TypeError) as excObj:
+def test_exactly() -> None:
+    with pytest.raises(TypeError):
         exactly('forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         exactly(-1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         exactly(1.0, 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         exactly(1)
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         exactly(1, '')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         exactly(1, '', '')
 
     assert exactly(1, 'cat') == 'cat{1}'
@@ -415,26 +421,26 @@ def test_exactly():
     assert exactly(0, 'cat') == 'cat{0}'
 
 
-def test_between():
-    with pytest.raises(TypeError) as excObj:
+def test_between() -> None:
+    with pytest.raises(TypeError):
         between('forty two', 1, 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         between(-1, 1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         between(1.0, 1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         between(1, 'forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         between(1, -1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         between(1, 1.0, 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         between(1, 2)
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         between(1, 2, '')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         between(1, 2, '', '')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         between(2, 1, '')
 
     assert between(1, 2, 'cat') == 'cat{1,2}'
@@ -443,18 +449,18 @@ def test_between():
     assert between(0, 0, 'cat') == 'cat{0,0}'
 
 
-def test_at_least():
-    with pytest.raises(TypeError) as excObj:
+def test_at_least() -> None:
+    with pytest.raises(TypeError):
         at_least('forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         at_least(-1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         at_least(1.0, 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         at_least(1)
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         at_least(1, '')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         at_least(1, '', '')
 
     assert at_least(1, 'cat') == 'cat{1,}'
@@ -463,18 +469,18 @@ def test_at_least():
     assert at_least(0, 'cat') == 'cat{0,}'
 
 
-def test_at_most():
-    with pytest.raises(TypeError) as excObj:
+def test_at_most() -> None:
+    with pytest.raises(TypeError):
         at_most('forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         at_most(-1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         at_most(1.0, 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         at_most(1)
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         at_most(1, '')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         at_most(1, '', '')
 
     assert at_most(1, 'cat') == 'cat{,1}'
@@ -483,55 +489,55 @@ def test_at_most():
     assert at_most(0, 'cat') == 'cat{,0}'
 
 
-def test_zero_or_more():
-    with pytest.raises(ValueError) as excObj:
+def test_zero_or_more() -> None:
+    with pytest.raises(ValueError):
         zero_or_more()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         zero_or_more('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         zero_or_more('', '')
 
     assert zero_or_more('x') == 'x*'
     assert zero_or_more('x', 'y') == 'xy*'
 
 
-def test_zero_or_more_lazy():
-    with pytest.raises(ValueError) as excObj:
+def test_zero_or_more_lazy() -> None:
+    with pytest.raises(ValueError):
         zero_or_more_lazy()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         zero_or_more_lazy('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         zero_or_more_lazy('', '')
 
     assert zero_or_more_lazy('x') == 'x*?'
     assert zero_or_more_lazy('x', 'y') == 'xy*?'
 
 
-def test_one_or_more():
-    with pytest.raises(ValueError) as excObj:
+def test_one_or_more() -> None:
+    with pytest.raises(ValueError):
         one_or_more()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         one_or_more('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         one_or_more('', '')
 
     assert one_or_more('x') == 'x+'
     assert one_or_more('x', 'y') == 'xy+'
 
 
-def test_one_or_more_lazy():
-    with pytest.raises(ValueError) as excObj:
+def test_one_or_more_lazy() -> None:
+    with pytest.raises(ValueError):
         one_or_more_lazy()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         one_or_more_lazy('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         one_or_more_lazy('', '')
 
     assert one_or_more_lazy('x') == 'x+?'
     assert one_or_more_lazy('x', 'y') == 'xy+?'
 
 
-def test_starts_with():
+def test_starts_with() -> None:
     assert starts_with('') == '^'
     assert starts_with('', '') == '^'
     assert starts_with() == '^'
@@ -539,7 +545,7 @@ def test_starts_with():
     assert starts_with('cat', 'dog', 'moose') == '^catdogmoose'
 
 
-def test_ends_with():
+def test_ends_with() -> None:
     assert ends_with('') == '$'
     assert ends_with('', '') == '$'
     assert ends_with() == '$'
@@ -547,7 +553,7 @@ def test_ends_with():
     assert ends_with('cat', 'dog', 'moose') == 'catdogmoose$'
 
 
-def test_starts_and_ends_with():
+def test_starts_and_ends_with() -> None:
     assert starts_and_ends_with('') == '^$'
     assert starts_and_ends_with('', '') == '^$'
     assert starts_and_ends_with() == '^$'
@@ -555,31 +561,31 @@ def test_starts_and_ends_with():
     assert starts_and_ends_with('cat', 'dog', 'moose') == '^catdogmoose$'
 
 
-def test_chars():
-    with pytest.raises(ValueError) as excObj:
+def test_chars() -> None:
+    with pytest.raises(ValueError):
         chars()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         chars('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         chars('', '')
 
     assert chars('a-z') == '[a-z]'
     assert chars('x', 'y') == '[xy]'
 
 
-def test_nonchars():
-    with pytest.raises(ValueError) as excObj:
+def test_nonchars() -> None:
+    with pytest.raises(ValueError):
         nonchars()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         nonchars('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         nonchars('', '')
 
     assert nonchars('a-z') == '[^a-z]'
     assert nonchars('x', 'y') == '[^xy]'
 
 
-def test_optional_group():
+def test_optional_group() -> None:
     assert optional_group() == '()?'
     assert optional_group('') == '()?'
     assert optional_group('c') == '(c)?'
@@ -588,7 +594,7 @@ def test_optional_group():
     assert group(optional_group('cat')) == '((cat)?)'
 
 
-def test_optional_noncap_group():
+def test_optional_noncap_group() -> None:
     assert optional_noncap_group() == '(?:)?'
     assert optional_noncap_group('') == '(?:)?'
     assert optional_noncap_group('c') == '(?:c)?'
@@ -597,7 +603,7 @@ def test_optional_noncap_group():
     assert group(optional_noncap_group('cat')) == '((?:cat)?)'
 
 
-def test_group_either():
+def test_group_either() -> None:
     assert group_either() == '()'
     assert group_either('') == '()'
     assert group_either('', '') == '()'
@@ -606,7 +612,7 @@ def test_group_either():
     assert group_either('cat', 'dog', 'moose') == '(cat|dog|moose)'
 
 
-def test_noncap_group_either():
+def test_noncap_group_either() -> None:
     assert noncap_group_either() == '(?:)'
     assert noncap_group_either('') == '(?:)'
     assert noncap_group_either('', '') == '(?:)'
@@ -615,12 +621,12 @@ def test_noncap_group_either():
     assert noncap_group_either('cat', 'dog', 'moose') == '(?:cat|dog|moose)'
 
 
-def test_group_exactly():
-    with pytest.raises(TypeError) as excObj:
+def test_group_exactly() -> None:
+    with pytest.raises(TypeError):
         group_exactly('forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         group_exactly(-1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         group_exactly(1.0, 'cat')
 
     assert group_exactly(1, 'cat') == '(cat){1}'
@@ -629,12 +635,12 @@ def test_group_exactly():
     assert group_exactly(0, 'cat') == '(cat){0}'
 
 
-def test_noncap_group_exactly():
-    with pytest.raises(TypeError) as excObj:
+def test_noncap_group_exactly() -> None:
+    with pytest.raises(TypeError):
         noncap_group_exactly('forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         noncap_group_exactly(-1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         noncap_group_exactly(1.0, 'cat')
 
     assert noncap_group_exactly(1, 'cat') == '(?:cat){1}'
@@ -643,22 +649,22 @@ def test_noncap_group_exactly():
     assert noncap_group_exactly(0, 'cat') == '(?:cat){0}'
 
 
-def test_group_between():
-    with pytest.raises(TypeError) as excObj:
+def test_group_between() -> None:
+    with pytest.raises(TypeError):
         group_between('forty two', 1, 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         group_between(-1, 1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         group_between(1.0, 1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         group_between(1, 'forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         group_between(1, -1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         group_between(1, 1.0, 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         group_between(2, 1, '')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         group_between(2, 1, '', '')
 
     assert group_between(1, 2) == '(){1,2}'
@@ -670,22 +676,22 @@ def test_group_between():
     assert group_between(0, 0, 'cat') == '(cat){0,0}'
 
 
-def test_noncap_group_between():
-    with pytest.raises(TypeError) as excObj:
+def test_noncap_group_between() -> None:
+    with pytest.raises(TypeError):
         noncap_group_between('forty two', 1, 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         noncap_group_between(-1, 1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         noncap_group_between(1.0, 1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         noncap_group_between(1, 'forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         noncap_group_between(1, -1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         noncap_group_between(1, 1.0, 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         noncap_group_between(2, 1, '')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         noncap_group_between(2, 1, '', '')
 
     assert noncap_group_between(1, 2) == '(?:){1,2}'
@@ -697,12 +703,12 @@ def test_noncap_group_between():
     assert noncap_group_between(0, 0, 'cat') == '(?:cat){0,0}'
 
 
-def test_group_at_least():
-    with pytest.raises(TypeError) as excObj:
+def test_group_at_least() -> None:
+    with pytest.raises(TypeError):
         group_at_least('forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         group_at_least(-1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         group_at_least(1.0, 'cat')
 
     assert group_at_least(1) == '(){1,}'
@@ -714,12 +720,12 @@ def test_group_at_least():
     assert group_at_least(0, 'cat') == '(cat){0,}'
 
 
-def test_noncap_group_at_least():
-    with pytest.raises(TypeError) as excObj:
+def test_noncap_group_at_least() -> None:
+    with pytest.raises(TypeError):
         noncap_group_at_least('forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         noncap_group_at_least(-1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         noncap_group_at_least(1.0, 'cat')
 
     assert noncap_group_at_least(1) == '(?:){1,}'
@@ -731,12 +737,12 @@ def test_noncap_group_at_least():
     assert noncap_group_at_least(0, 'cat') == '(?:cat){0,}'
 
 
-def test_group_at_most():
-    with pytest.raises(TypeError) as excObj:
+def test_group_at_most() -> None:
+    with pytest.raises(TypeError):
         group_at_most('forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         group_at_most(-1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         group_at_most(1.0, 'cat')
 
     assert group_at_most(1) == '(){,1}'
@@ -748,12 +754,12 @@ def test_group_at_most():
     assert group_at_most(0, 'cat') == '(cat){,0}'
 
 
-def test_noncap_group_at_most():
-    with pytest.raises(TypeError) as excObj:
+def test_noncap_group_at_most() -> None:
+    with pytest.raises(TypeError):
         noncap_group_at_most('forty two', 'cat')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         noncap_group_at_most(-1, 'cat')
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         noncap_group_at_most(1.0, 'cat')
 
     assert noncap_group_at_most(1) == '(?:){,1}'
@@ -765,7 +771,7 @@ def test_noncap_group_at_most():
     assert noncap_group_at_most(0, 'cat') == '(?:cat){,0}'
 
 
-def test_zero_or_more_group():
+def test_zero_or_more_group() -> None:
     assert zero_or_more_group() == '()*'
     assert zero_or_more_group('') == '()*'
     assert zero_or_more_group('', '') == '()*'
@@ -773,7 +779,7 @@ def test_zero_or_more_group():
     assert zero_or_more_group('x', 'y') == '(xy)*'
 
 
-def test_zero_or_more_noncap_group():
+def test_zero_or_more_noncap_group() -> None:
     assert zero_or_more_noncap_group() == '(?:)*'
     assert zero_or_more_noncap_group('') == '(?:)*'
     assert zero_or_more_noncap_group('', '') == '(?:)*'
@@ -781,7 +787,7 @@ def test_zero_or_more_noncap_group():
     assert zero_or_more_noncap_group('x', 'y') == '(?:xy)*'
 
 
-def test_zero_or_more_lazy_group():
+def test_zero_or_more_lazy_group() -> None:
     assert zero_or_more_lazy_group() == '()*?'
     assert zero_or_more_lazy_group('') == '()*?'
     assert zero_or_more_lazy_group('', '') == '()*?'
@@ -789,7 +795,7 @@ def test_zero_or_more_lazy_group():
     assert zero_or_more_lazy_group('x', 'y') == '(xy)*?'
 
 
-def test_zero_or_more_lazy_noncap_group():
+def test_zero_or_more_lazy_noncap_group() -> None:
     assert zero_or_more_lazy_noncap_group() == '(?:)*?'
     assert zero_or_more_lazy_noncap_group('') == '(?:)*?'
     assert zero_or_more_lazy_noncap_group('', '') == '(?:)*?'
@@ -797,7 +803,7 @@ def test_zero_or_more_lazy_noncap_group():
     assert zero_or_more_lazy_noncap_group('x', 'y') == '(?:xy)*?'
 
 
-def test_one_or_more_group():
+def test_one_or_more_group() -> None:
     assert one_or_more_group() == '()+'
     assert one_or_more_group('') == '()+'
     assert one_or_more_group('', '') == '()+'
@@ -805,7 +811,7 @@ def test_one_or_more_group():
     assert one_or_more_group('x', 'y') == '(xy)+'
 
 
-def test_one_or_more_noncap_group():
+def test_one_or_more_noncap_group() -> None:
     assert one_or_more_noncap_group() == '(?:)+'
     assert one_or_more_noncap_group('') == '(?:)+'
     assert one_or_more_noncap_group('', '') == '(?:)+'
@@ -813,7 +819,7 @@ def test_one_or_more_noncap_group():
     assert one_or_more_noncap_group('x', 'y') == '(?:xy)+'
 
 
-def test_one_or_more_lazy_group():
+def test_one_or_more_lazy_group() -> None:
     assert one_or_more_lazy_group() == '()+?'
     assert one_or_more_lazy_group('') == '()+?'
     assert one_or_more_lazy_group('', '') == '()+?'
@@ -821,7 +827,7 @@ def test_one_or_more_lazy_group():
     assert one_or_more_lazy_group('x', 'y') == '(xy)+?'
 
 
-def test_one_or_more_lazy_noncap_group():
+def test_one_or_more_lazy_noncap_group() -> None:
     assert one_or_more_lazy_noncap_group() == '(?:)+?'
     assert one_or_more_lazy_noncap_group('') == '(?:)+?'
     assert one_or_more_lazy_noncap_group('', '') == '(?:)+?'
@@ -829,60 +835,60 @@ def test_one_or_more_lazy_noncap_group():
     assert one_or_more_lazy_noncap_group('x', 'y') == '(?:xy)+?'
 
 
-def test_group_chars():
-    with pytest.raises(ValueError) as excObj:
+def test_group_chars() -> None:
+    with pytest.raises(ValueError):
         group_chars()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         group_chars('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         group_chars('', '')
 
     assert group_chars('a-z') == '([a-z])'
     assert group_chars('x', 'y') == '([xy])'
 
 
-def test_noncap_group_chars():
-    with pytest.raises(ValueError) as excObj:
+def test_noncap_group_chars() -> None:
+    with pytest.raises(ValueError):
         noncap_group_chars()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         noncap_group_chars('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         noncap_group_chars('', '')
 
     assert noncap_group_chars('a-z') == '(?:[a-z])'
     assert noncap_group_chars('x', 'y') == '(?:[xy])'
 
 
-def test_group_nonchars():
-    with pytest.raises(ValueError) as excObj:
+def test_group_nonchars() -> None:
+    with pytest.raises(ValueError):
         group_nonchars()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         group_nonchars('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         group_nonchars('', '')
 
     assert group_nonchars('a-z') == '([^a-z])'
     assert group_nonchars('x', 'y') == '([^xy])'
 
 
-def test_noncap_group_nonchars():
-    with pytest.raises(ValueError) as excObj:
+def test_noncap_group_nonchars() -> None:
+    with pytest.raises(ValueError):
         noncap_group_nonchars()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         noncap_group_nonchars('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         noncap_group_nonchars('', '')
 
     assert noncap_group_nonchars('a-z') == '(?:[^a-z])'
     assert noncap_group_nonchars('x', 'y') == '(?:[^xy])'
 
 
-def test_back_reference():
-    with pytest.raises(TypeError) as excObj:
+def test_back_reference() -> None:
+    with pytest.raises(TypeError):
         back_reference('one')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         back_reference(-1)
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         back_reference(-0)
 
     for i in range(1, 20):
@@ -898,11 +904,11 @@ def test_back_reference():
     assert back_reference(8) == BACK_8
     assert back_reference(9) == BACK_9
 
-    with pytest.raises(TypeError) as excObj:
+    with pytest.raises(TypeError):
         back_ref('one')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         back_ref(-1)
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         back_ref(-0)
 
     for i in range(1, 20):
@@ -912,7 +918,7 @@ def test_back_reference():
 # Taking out these ASCII_* tests since the constants have been
 # removed for now (and probably permanently.)
 """
-def test_ascii_letter_class():
+def test_ascii_letter_class() -> None:
     assert compile(ASCII_LETTER).match('') is None
     for i in range(65, 91):  # A-Z
         assert compile(ASCII_LETTER).match(chr(i))
@@ -926,7 +932,7 @@ def test_ascii_letter_class():
         assert compile(ASCII_LETTER).match(chr(i)) is None
 
 
-def test_ascii_nonletter_class():
+def test_ascii_nonletter_class() -> None:
     assert compile(ASCII_NONLETTER).match('') is None
     for i in range(65, 91):  # A-Z
         assert compile(ASCII_NONLETTER).match(chr(i)) is None
@@ -940,7 +946,7 @@ def test_ascii_nonletter_class():
         assert compile(ASCII_NONLETTER).match(chr(i))
 
 
-def test_ascii_uppercase_class():
+def test_ascii_uppercase_class() -> None:
     assert compile(ASCII_UPPERCASE).match('') is None
     for i in range(65, 91):  # A-Z
         assert compile(ASCII_UPPERCASE).match(chr(i))
@@ -954,7 +960,7 @@ def test_ascii_uppercase_class():
         assert compile(ASCII_UPPERCASE).match(chr(i)) is None
 
 
-def test_ascii_nonuppercase_class():
+def test_ascii_nonuppercase_class() -> None:
     assert compile(ASCII_NONUPPERCASE).match('') is None
     for i in range(65, 91):  # A-Z
         assert compile(ASCII_NONUPPERCASE).match(chr(i)) is None
@@ -968,7 +974,7 @@ def test_ascii_nonuppercase_class():
         assert compile(ASCII_NONUPPERCASE).match(chr(i))
 
 
-def test_ascii_lowercase_class():
+def test_ascii_lowercase_class() -> None:
     assert compile(ASCII_LOWERCASE).match('') is None
     for i in range(65, 91):  # A-Z
         assert compile(ASCII_LOWERCASE).match(chr(i)) is None
@@ -982,7 +988,7 @@ def test_ascii_lowercase_class():
         assert compile(ASCII_LOWERCASE).match(chr(i)) is None
 
 
-def test_ascii_nonlowercase_class():
+def test_ascii_nonlowercase_class() -> None:
     assert compile(ASCII_NONLOWERCASE).match('') is None
     for i in range(65, 91):  # A-Z
         assert compile(ASCII_NONLOWERCASE).match(chr(i))
@@ -996,7 +1002,7 @@ def test_ascii_nonlowercase_class():
         assert compile(ASCII_NONLOWERCASE).match(chr(i))
 
 
-def test_ascii_alphanumeric():
+def test_ascii_alphanumeric() -> None:
     assert compile(ASCII_ALPHANUMERIC).match('') is None
     for i in range(65, 91):  # A-Z
         assert compile(ASCII_ALPHANUMERIC).match(chr(i))
@@ -1010,7 +1016,7 @@ def test_ascii_alphanumeric():
         assert compile(ASCII_ALPHANUMERIC).match(chr(i)) is None
 
 
-def test_ascii_nonalphanumeric_class():
+def test_ascii_nonalphanumeric_class() -> None:
     assert compile(ASCII_NONALPHANUMERIC).match('') is None
     for i in range(65, 91):  # A-Z
         assert compile(ASCII_NONALPHANUMERIC).match(chr(i)) is None
@@ -1024,7 +1030,7 @@ def test_ascii_nonalphanumeric_class():
         assert compile(ASCII_NONALPHANUMERIC).match(chr(i))
 
 
-def test_ascii_numeric():
+def test_ascii_numeric() -> None:
     # TODO - test that this doesn't match numeric characters outside of 0-9.
     assert compile(ASCII_NUMERIC).match('') is None
     for i in range(65, 91):  # A-Z
@@ -1039,7 +1045,7 @@ def test_ascii_numeric():
         assert compile(ASCII_NUMERIC).match(chr(i)) is None
 
 
-def test_ascii_nonnumeric():
+def test_ascii_nonnumeric() -> None:
     assert compile(ASCII_NONNUMERIC).match('') is None
     for i in range(65, 91):  # A-Z
         assert compile(ASCII_NONNUMERIC).match(chr(i))
@@ -1054,7 +1060,7 @@ def test_ascii_nonnumeric():
 """
 
 
-def test_number_pattern():
+def test_number_pattern() -> None:
     assert compile(NUMBER).match('') is None
     for i in range(65, 91):  # A-Z
         assert compile(NUMBER).match(chr(i)) is None
@@ -1080,7 +1086,7 @@ def test_number_pattern():
     assert compile(NUMBER).match('-1,234.00')
 
 
-def test_euro_number_pattern():
+def test_euro_number_pattern() -> None:
     for i in range(65, 91):  # A-Z
         assert compile(EURO_NUMBER).match(chr(i)) is None
     for i in range(97, 123):  # a-z
@@ -1105,7 +1111,7 @@ def test_euro_number_pattern():
     assert compile(EURO_NUMBER).match('-1.234,00')
 
 
-def test_hexadecimal():
+def test_hexadecimal() -> None:
     assert compile(HEXADECIMAL).match('') is None
     for i in range(65, 71):  # A-F
         assert compile(HEXADECIMAL).match(chr(i))
@@ -1123,7 +1129,7 @@ def test_hexadecimal():
         assert compile(HEXADECIMAL).match(chr(i)) is None
 
 
-def test_nonhexadecimal():
+def test_nonhexadecimal() -> None:
     assert compile(NONHEXADECIMAL).match('') is None
     for i in range(65, 71):  # A-F
         assert compile(NONHEXADECIMAL).match(chr(i)) is None
@@ -1141,7 +1147,7 @@ def test_nonhexadecimal():
         assert compile(NONHEXADECIMAL).match(chr(i))
 
 
-def test_hexadecimal_number():
+def test_hexadecimal_number() -> None:
     assert compile(HEXADECIMAL_NUMBER).match('') is None
     assert compile(HEXADECIMAL_NUMBER).match('100')
     assert compile(HEXADECIMAL_NUMBER).match('FF')
@@ -1154,107 +1160,107 @@ def test_hexadecimal_number():
     assert compile(HEXADECIMAL_NUMBER).match('0X1234567890ABCDEF')
 
 
-def test_letter_character_class():
+def test_letter_character_class() -> None:
     assert compile(LETTER).match('') is None
     # for i in range(0, sys.maxunicode + 1):
-    for i in range(0, 1000):
+    for i in range(1000):
         if chr(i).isalpha():
             assert compile(LETTER).match(chr(i))
         else:
             assert compile(LETTER).match(chr(i)) is None
 
 
-def test_nonletter_character_class():
+def test_nonletter_character_class() -> None:
     assert compile(NONLETTER).match('') is None
     # for i in range(0, sys.maxunicode + 1):
-    for i in range(0, 1000):
+    for i in range(1000):
         if not chr(i).isalpha():
             assert compile(NONLETTER).match(chr(i))
         else:
             assert compile(NONLETTER).match(chr(i)) is None
 
 
-def test_lowercase_character_class():
+def test_lowercase_character_class() -> None:
     assert compile(LOWERCASE).match('') is None
     # for i in range(0, sys.maxunicode + 1):
-    for i in range(0, 1000):
+    for i in range(1000):
         if chr(i).islower():
             assert compile(LOWERCASE).match(chr(i))
         else:
             assert compile(LOWERCASE).match(chr(i)) is None
 
 
-def test_nonlowercase_character_class():
+def test_nonlowercase_character_class() -> None:
     assert compile(NONLOWERCASE).match('') is None
     # for i in range(0, sys.maxunicode + 1):
-    for i in range(0, 1000):
+    for i in range(1000):
         if not chr(i).islower():
             assert compile(NONLOWERCASE).match(chr(i))
         else:
             assert compile(NONLOWERCASE).match(chr(i)) is None
 
 
-def test_uppercase_character_class():
+def test_uppercase_character_class() -> None:
     assert compile(UPPERCASE).match('') is None
     # for i in range(0, sys.maxunicode + 1):
-    for i in range(0, 1000):
+    for i in range(1000):
         if chr(i).isupper():
             assert compile(UPPERCASE).match(chr(i))
         else:
             assert compile(UPPERCASE).match(chr(i)) is None
 
 
-def test_nonuppercase_character_class():
+def test_nonuppercase_character_class() -> None:
     assert compile(NONUPPERCASE).match('') is None
     # for i in range(0, sys.maxunicode + 1):
-    for i in range(0, 1000):
+    for i in range(1000):
         if not chr(i).isupper():
             assert compile(NONUPPERCASE).match(chr(i))
         else:
             assert compile(NONUPPERCASE).match(chr(i)) is None
 
 
-def test_alphanumeric_character_class():
+def test_alphanumeric_character_class() -> None:
     assert compile(ALPHANUMERIC).match('') is None
     # for i in range(0, sys.maxunicode + 1):
-    for i in range(0, 1000):
+    for i in range(1000):
         if chr(i).isalnum():
             assert compile(ALPHANUMERIC).match(chr(i))
         else:
             assert compile(ALPHANUMERIC).match(chr(i)) is None
 
 
-def test_nonalphanumeric_character_class():
+def test_nonalphanumeric_character_class() -> None:
     assert compile(NONALPHANUMERIC).match('') is None
     # for i in range(0, sys.maxunicode + 1):
-    for i in range(0, 1000):
+    for i in range(1000):
         if not chr(i).isalnum():
             assert compile(NONALPHANUMERIC).match(chr(i))
         else:
             assert compile(NONALPHANUMERIC).match(chr(i)) is None
 
 
-def test_numeric_character_class():
+def test_numeric_character_class() -> None:
     assert compile(NUMERIC).match('') is None
     # for i in range(0, sys.maxunicode + 1):
-    for i in range(0, 1000):
+    for i in range(1000):
         if chr(i).isnumeric():
             assert compile(NUMERIC).match(chr(i))
         else:
             assert compile(NUMERIC).match(chr(i)) is None
 
 
-def test_nonnumeric_character_class():
+def test_nonnumeric_character_class() -> None:
     assert compile(NONNUMERIC).match('') is None
     # for i in range(0, sys.maxunicode + 1):
-    for i in range(0, 1000):
+    for i in range(1000):
         if not chr(i).isnumeric():
             assert compile(NONNUMERIC).match(chr(i))
         else:
             assert compile(NONNUMERIC).match(chr(i)) is None
 
 
-def test_constants():
+def test_constants() -> None:
     assert DIGIT == r'\d'
     assert WORD == r'\w'
     assert WHITESPACE == r'\s'
@@ -1320,7 +1326,7 @@ def test_constants():
     # Built-in Humre Patterns:
     assert ANYTHING == '.*?'
     assert EVERYTHING == '.*'
-    #assert GREEDY_SOMETHING == '.+'
+    # assert GREEDY_SOMETHING == '.+'
     assert SOMETHING == '.+?'
     assert ANYCHAR == '.'
 
@@ -1356,14 +1362,14 @@ def test_constants():
     assert BACK_9 == r'\9'
 
 
-def test_inline_flags():
-    with pytest.raises(TypeError) as excObj:
+def test_inline_flags() -> None:
+    with pytest.raises(TypeError):
         inline_flag(42, 'foo', 'bar')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         inline_flag('au', 'foo', 'bar')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         inline_flag('ua', 'foo', 'bar')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         inline_flag('z', 'foo', 'bar')
 
     assert inline_flag('a', 'foo', 'bar') == '(?a:foobar)'
@@ -1379,7 +1385,7 @@ def test_inline_flags():
     # TODO - test that exceptions get raised
 
 
-def test_atomic_group():
+def test_atomic_group() -> None:
     assert atomic_group() == '(?>)'
     assert atomic_group('') == '(?>)'
     assert atomic_group('', '') == '(?>)'
@@ -1387,36 +1393,36 @@ def test_atomic_group():
     assert atomic_group('x', 'y') == '(?>xy)'
 
 
-def test_zero_or_more_possessive():
-    with pytest.raises(ValueError) as excObj:
+def test_zero_or_more_possessive() -> None:
+    with pytest.raises(ValueError):
         zero_or_more_possessive()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         zero_or_more_possessive('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         zero_or_more_possessive('', '')
 
     assert zero_or_more_possessive('x') == 'x*+'
     assert zero_or_more_possessive('x', 'y') == 'xy*+'
 
 
-def test_one_or_more_possessive():
-    with pytest.raises(ValueError) as excObj:
+def test_one_or_more_possessive() -> None:
+    with pytest.raises(ValueError):
         one_or_more_possessive()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         one_or_more_possessive('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         one_or_more_possessive('', '')
 
     assert one_or_more_possessive('x') == 'x++'
     assert one_or_more_possessive('x', 'y') == 'xy++'
 
 
-def test_optional_possessive():
-    with pytest.raises(ValueError) as excObj:
+def test_optional_possessive() -> None:
+    with pytest.raises(ValueError):
         optional_possessive()
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         optional_possessive('')
-    with pytest.raises(ValueError) as excObj:
+    with pytest.raises(ValueError):
         optional_possessive('', '')
 
     assert optional_possessive('x') == 'x?+'
